chore: move RNSScreenStackHeaderSubviewType enum to RNSEnums & add conversion method
