+++ ios/RNSConvert.h
@@ -12,6 +12,9 @@
 + (RNSScreenStackAnimation)RNSScreenStackAnimationFromCppEquivalent:
     (facebook::react::RNSScreenStackAnimation)stackAnimation;
 
++ (RNSScreenStackHeaderSubviewType)RNSScreenStackHeaderSubviewTypeFromCppEquivalent:
+    (facebook::react::RNSScreenStackHeaderSubviewType)subviewType;
+
 @end
 
 #endif /* RNSConvert_h */
